Improve version handling (#23)
* Add a script to minify the json

This is in the setup.py but it is useful to be able to run it on its own

* Update pymctranslate

* Improve version handling

Remove min_version and max_version. These behaviour of these properties was confusing.
Add (min/max/max_known) semantic_version and block_version properties.
These are clearer what they do.
max_known indicates a real version to use as an upper bound. max is generally not a real version and exists purely to catch future versions.

* Remove unused class

* Reformat

* Reformat

* Expose get_max_known_block_version to C++